fixed a bug in uart module and in angle to compare conversion

+++ components/uart/include/uart_comm.h
@@ -10,23 +10,23 @@
 #define PACKET_HEADER   0xAA
 
 typedef struct __attribute__((packed)){
-    float x;
-    float y;
+    int32_t x;
+    int32_t y;
 } ball_pos_t;
 
 typedef struct __attribute__((packed)){
-    int32_t servo_id;   // 1, 2 o 3
-    int32_t angle;      // Angolo in gradi (es. -30 a +30)
+    int32_t servo_id;   
+    int32_t angle;      
 } servo_row_t;
 
 typedef struct __attribute__((packed)){
-    uint8_t header;         // Byte 0: 0xAA
-    uint8_t cmd_type;       // Byte 1: 0x01 o 0x02
-    union {                 // Byte 2-9: 8 byte interpretati diversamente
+    uint8_t header;         
+    uint8_t cmd_type;       
+    union {                 
         ball_pos_t tracking;
         servo_row_t servo;
     } payload;
-    uint8_t checksum;       // Byte 10: XOR byte 0-9
+    uint8_t checksum;
 } uart_packet_t;
 
 extern QueueHandle_t ball_pos_queue;
